Have LdapPasswordManagerService.changeUserPassword change pwd for all servers
Separate the authentication of the old password from the password change,
so that it will change the password for all servers.  The former behavior
was to authenticate for each server and only change the ones where the
old password was valid.  Now, it will change all passwords if any of them
matches the old password.