LdapPasswordManagerService.getUserSecurityChallenge - Don't stop at null challenge.
The original behavior was to stop and return the security challenge from
the first server where the username exists, whether or not it was set.  The
new behavior is to keep going if the security challenge is null and only
return null when there is no server with a non-null security challenge.